Split roles.rs into separate files
