--- daphne/src/roles.rs
@@ -1,1337 +1,0 @@
-// Copyright (c) 2022 Cloudflare, Inc. All rights reserved.
-// SPDX-License-Identifier: BSD-3-Clause
-
-//! Trait definitions for Daphne backends.
-
-use crate::{
-    audit_log::{AggregationJobAuditAction, AuditLog},
-    constants::DapMediaType,
-    fatal_error,
-    hpke::{HpkeConfig, HpkeDecrypter},
-    messages::{
-        constant_time_eq, decode_base64url, AggregateShare, AggregateShareReq,
-        AggregationJobContinueReq, AggregationJobInitReq, AggregationJobResp, BatchId,
-        BatchSelector, Collection, CollectionJobId, CollectionReq, HpkeConfigList, Interval,
-        PartialBatchSelector, Query, Report, ReportMetadata, TaskId, Time, TransitionFailure,
-    },
-    metrics::{DaphneMetrics, DaphneRequestType},
-    taskprov,
-    vdaf::{EarlyReportStateConsumed, EarlyReportStateInitialized},
-    DapAbort, DapAggregateShare, DapCollectJob, DapError, DapGlobalConfig, DapHelperState,
-    DapHelperTransition, DapLeaderProcessTelemetry, DapLeaderTransition, DapOutputShare,
-    DapQueryConfig, DapRequest, DapResource, DapResponse, DapTaskConfig, DapVersion,
-    MetaAggregationJobId,
-};
-use async_trait::async_trait;
-use futures::TryFutureExt;
-use prio::codec::{Decode, Encode, ParameterizedDecode, ParameterizedEncode};
-use std::borrow::Cow;
-use std::collections::HashMap;
-use tracing::{debug, error, warn};
-use url::Url;
-
-/// A party in the DAP protocol who is authorized to send requests to another party.
-#[async_trait(?Send)]
-pub trait DapAuthorizedSender<S> {
-    /// Add authorization to an outbound DAP request with the given task ID, media type, and payload.
-    async fn authorize(
-        &self,
-        task_id: &TaskId,
-        media_type: &DapMediaType,
-        payload: &[u8],
-    ) -> Result<S, DapError>;
-}
-
-/// Report initializer. Used by a DAP Aggregator [`DapAggregator`] when initializing an aggregation
-/// job.
-#[async_trait(?Send)]
-pub trait DapReportInitializer {
-    /// Initialize a sequence of reports that are in the "consumed" state by performing the early
-    /// validation steps (check if the report was replayed, belongs to a batch that has been
-    /// collected) and initializing VDAF preparation.
-    async fn initialize_reports<'req>(
-        &self,
-        is_leader: bool,
-        task_id: &TaskId,
-        task_config: &DapTaskConfig,
-        part_batch_sel: &PartialBatchSelector,
-        consumed_reports: Vec<EarlyReportStateConsumed<'req>>,
-    ) -> Result<Vec<EarlyReportStateInitialized<'req>>, DapError>;
-}
-
-/// DAP Aggregator functionality.
-#[async_trait(?Send)]
-pub trait DapAggregator<S>: HpkeDecrypter + DapReportInitializer + Sized {
-    /// A refernce to a task configuration stored by the Aggregator.
-    type WrappedDapTaskConfig<'a>: AsRef<DapTaskConfig>;
-
-    /// Decide whether the given DAP request is authorized.
-    ///
-    /// If the return value is `None`, then the request is authorized. If the return value is
-    /// `Some(reason)`, then the request is denied and `reason` conveys details about how the
-    /// decision was reached.
-    async fn unauthorized_reason(&self, req: &DapRequest<S>) -> Result<Option<String>, DapError>;
-
-    /// Look up the DAP global configuration.
-    fn get_global_config(&self) -> &DapGlobalConfig;
-
-    /// taskprov: The VDAF verification key initializer. Used to derive the VDAF verify key for all
-    /// tasks configured by this extension.
-    fn taskprov_vdaf_verify_key_init(&self) -> Option<&[u8; 32]>;
-
-    /// taskprov: The Collector's HPKE configuration used for all tasks configured by this
-    /// extension.
-    fn taskprov_collector_hpke_config(&self) -> Option<&HpkeConfig>;
-
-    /// taskprov: Decide whether to opt-in or out-out of a task provisioned via taskprov.
-    ///
-    /// If the return value is `None`, then the decision is to opt-in. If the return value is
-    /// `Some(reason)`, then the decision is to opt-out; `reason` conveys details about how the
-    /// decision was reached (e.g.., the minimum batch size is too smal).
-    fn taskprov_opt_out_reason(
-        &self,
-        task_config: &DapTaskConfig,
-    ) -> Result<Option<String>, DapError>;
-
-    /// taskprov: Configure a task. This is called after opting in. If successful, the next call to
-    /// `get_task_config_for()` will return the configure task. Otherwise this call will return
-    /// nothing.
-    async fn taskprov_put(
-        &self,
-        req: &DapRequest<S>,
-        task_config: DapTaskConfig,
-    ) -> Result<(), DapError>;
-
-    /// Look up the DAP task configuration for the given task ID.
-    async fn get_task_config_for<'req>(
-        &self,
-        task_id: Cow<'req, TaskId>,
-    ) -> Result<Option<Self::WrappedDapTaskConfig<'req>>, DapError>;
-
-    /// Get the current time (number of seconds since the beginning of UNIX time).
-    fn get_current_time(&self) -> Time;
-
-    /// Check whether the batch determined by the collect request would overlap with a previous
-    /// batch.
-    async fn is_batch_overlapping(
-        &self,
-        task_id: &TaskId,
-        batch_sel: &BatchSelector,
-    ) -> Result<bool, DapError>;
-
-    /// Check whether the given batch ID has been observed before. This is called by the Leader
-    /// (resp. Helper) in response to a CollectReq (resp. AggregateShareReq) for fixed-size tasks.
-    async fn batch_exists(&self, task_id: &TaskId, batch_id: &BatchId) -> Result<bool, DapError>;
-
-    /// Store a set of output shares.
-    async fn put_out_shares(
-        &self,
-        task_id: &TaskId,
-        part_batch_sel: &PartialBatchSelector,
-        out_shares: Vec<DapOutputShare>,
-    ) -> Result<(), DapError>;
-
-    /// Fetch the aggregate share for the given batch.
-    async fn get_agg_share(
-        &self,
-        task_id: &TaskId,
-        batch_sel: &BatchSelector,
-    ) -> Result<DapAggregateShare, DapError>;
-
-    /// Mark a batch as collected.
-    async fn mark_collected(
-        &self,
-        task_id: &TaskId,
-        batch_sel: &BatchSelector,
-    ) -> Result<(), DapError>;
-
-    /// Handle request for the Aggregator's HPKE configuration.
-    async fn handle_hpke_config_req(&self, req: &DapRequest<S>) -> Result<DapResponse, DapAbort> {
-        // Check whether the DAP version indicated by the sender is supported.
-        if req.version == DapVersion::Unknown {
-            return Err(DapAbort::version_unknown());
-        }
-
-        let metrics = self.metrics().with_host(req.host());
-
-        // Parse the task ID from the query string, ensuring that it is the only query parameter.
-        let mut id = None;
-        for (k, v) in req.url.query_pairs() {
-            if k != "task_id" {
-                return Err(DapAbort::BadRequest("unexpected query parameter".into()));
-            }
-
-            let bytes = decode_base64url(v.as_bytes()).ok_or(DapAbort::BadRequest(
-                "failed to parse query parameter as URL-safe Base64".into(),
-            ))?;
-
-            id = Some(TaskId(bytes))
-        }
-
-        let hpke_config = self.get_hpke_config_for(req.version, id.as_ref()).await?;
-
-        if let Some(task_id) = id {
-            let task_config = self
-                .get_task_config_for(Cow::Owned(task_id))
-                .await?
-                .ok_or(DapAbort::UnrecognizedTask)?;
-
-            // Check whether the DAP version in the request matches the task config.
-            if task_config.as_ref().version != req.version {
-                return Err(DapAbort::version_mismatch(
-                    req.version,
-                    task_config.as_ref().version,
-                ));
-            }
-        }
-
-        let payload = match req.version {
-            DapVersion::Draft02 => hpke_config.as_ref().get_encoded(),
-            DapVersion::Draft05 => {
-                let hpke_config_list = HpkeConfigList {
-                    hpke_configs: vec![hpke_config.as_ref().clone()],
-                };
-                hpke_config_list.get_encoded()
-            }
-            // This is just to keep the compiler happy as we excluded DapVersion::Unknown by
-            // aborting at the top of the function.
-            _ => unreachable!("unhandled version {:?}", req.version),
-        };
-
-        metrics.inbound_req_inc(DaphneRequestType::HpkeConfig);
-        Ok(DapResponse {
-            version: req.version,
-            media_type: DapMediaType::HpkeConfigList,
-            payload,
-        })
-    }
-
-    async fn current_batch(&self, task_id: &TaskId) -> Result<BatchId, DapError>;
-
-    /// Access the Prometheus metrics.
-    fn metrics(&self) -> &DaphneMetrics;
-
-    /// Access the audit log.
-    fn audit_log(&self) -> &dyn AuditLog;
-}
-
-macro_rules! leader_post {
-    (
-        $role:expr,
-        $task_id:expr,
-        $task_config:expr,
-        $path:expr,
-        $req_media_type:expr,
-        $resp_media_type:expr,
-        $resource:expr,
-        $req_data:expr,
-        $is_put:expr
-    ) => {{
-        let url = $task_config
-            .helper_url
-            .join($path)
-            .map_err(|e| fatal_error!(err = e))?;
-
-        let req = DapRequest {
-            version: $task_config.version,
-            media_type: $req_media_type,
-            task_id: Some($task_id.clone()),
-            resource: $resource,
-            payload: $req_data,
-            url,
-            sender_auth: Some(
-                $role
-                    .authorize(&$task_id, &$req_media_type, &$req_data)
-                    .await?,
-            ),
-            taskprov: None,
-        };
-
-        let resp = if $is_put {
-            $role.send_http_put(req).await?
-        } else {
-            $role.send_http_post(req).await?
-        };
-
-        check_response_content_type(&resp, $resp_media_type)?;
-        resp
-    }};
-}
-
-/// DAP Leader functionality.
-#[async_trait(?Send)]
-pub trait DapLeader<S>: DapAuthorizedSender<S> + DapAggregator<S> {
-    /// Data type used to guide selection of a set of reports for aggregation.
-    type ReportSelector;
-
-    /// Store a report for use later on.
-    async fn put_report(&self, report: &Report, task_id: &TaskId) -> Result<(), DapError>;
-
-    /// Fetch a sequence of reports to aggregate, grouped by task ID, then by partial batch
-    /// selector. The reports returned are removed from persistent storage.
-    async fn get_reports(
-        &self,
-        selector: &Self::ReportSelector,
-    ) -> Result<HashMap<TaskId, HashMap<PartialBatchSelector, Vec<Report>>>, DapError>;
-
-    /// Create a collect job.
-    //
-    // TODO spec: Figure out if the hostname for the collect URI needs to match the Leader.
-    async fn init_collect_job(
-        &self,
-        task_id: &TaskId,
-        collect_job_id: &Option<CollectionJobId>,
-        collect_req: &CollectionReq,
-    ) -> Result<Url, DapError>;
-
-    /// Check the status of a collect job.
-    async fn poll_collect_job(
-        &self,
-        task_id: &TaskId,
-        collect_id: &CollectionJobId,
-    ) -> Result<DapCollectJob, DapError>;
-
-    /// Fetch the current collect job queue. The result is the sequence of collect ID and request
-    /// pairs, in order of priority.
-    async fn get_pending_collect_jobs(
-        &self,
-    ) -> Result<Vec<(TaskId, CollectionJobId, CollectionReq)>, DapError>;
-
-    /// Complete a collect job by assigning it the completed [`CollectResp`](crate::messages::Collection).
-    async fn finish_collect_job(
-        &self,
-        task_id: &TaskId,
-        collect_id: &CollectionJobId,
-        collect_resp: &Collection,
-    ) -> Result<(), DapError>;
-
-    /// Send an HTTP POST request.
-    async fn send_http_post(&self, req: DapRequest<S>) -> Result<DapResponse, DapError>;
-
-    /// Send an HTTP PUT request.
-    async fn send_http_put(&self, req: DapRequest<S>) -> Result<DapResponse, DapError>;
-
-    /// Handle a report from a Client.
-    async fn handle_upload_req(&self, req: &DapRequest<S>) -> Result<(), DapAbort> {
-        let metrics = self.metrics().with_host(req.host());
-        let task_id = req.task_id()?;
-        debug!("upload for task {task_id}");
-
-        // Check whether the DAP version indicated by the sender is supported.
-        if req.version == DapVersion::Unknown {
-            return Err(DapAbort::version_unknown());
-        }
-
-        check_request_content_type(req, DapMediaType::Report)?;
-
-        let report = Report::get_decoded_with_param(&req.version, req.payload.as_ref())
-            .map_err(|e| DapAbort::from_codec_error(e, task_id.clone()))?;
-        debug!("report id is {}", report.report_metadata.id);
-
-        resolve_taskprov(self, task_id, req, Some(&report.report_metadata)).await?;
-        let task_config = self
-            .get_task_config_for(Cow::Borrowed(task_id))
-            .await?
-            .ok_or(DapAbort::UnrecognizedTask)?;
-
-        // Check whether the DAP version in the request matches the task config.
-        if task_config.as_ref().version != req.version {
-            return Err(DapAbort::version_mismatch(
-                req.version,
-                task_config.as_ref().version,
-            ));
-        }
-
-        if report.encrypted_input_shares.len() != 2 {
-            // TODO spec: Decide if this behavior should be specified.
-            return Err(DapAbort::UnrecognizedMessage {
-                detail: format!(
-                    "expected exactly two encrypted input shares; got {}",
-                    report.encrypted_input_shares.len()
-                ),
-                task_id: Some(task_id.clone()),
-            });
-        }
-
-        // Check that the indicated HpkeConfig is present.
-        //
-        // TODO spec: It's not clear if this behavior is MUST, SHOULD, or MAY.
-        if !self
-            .can_hpke_decrypt(req.task_id()?, report.encrypted_input_shares[0].config_id)
-            .await?
-        {
-            return Err(DapAbort::ReportRejected {
-                detail: "No current HPKE configuration matches the indicated ID.".into(),
-            });
-        }
-
-        // Check that the task has not expired.
-        if report.report_metadata.time >= task_config.as_ref().expiration {
-            return Err(DapAbort::ReportTooLate);
-        }
-
-        // Store the report for future processing. At this point, the report may be rejected if
-        // the Leader detects that the report was replayed or pertains to a batch that has already
-        // been collected.
-        self.put_report(&report, req.task_id()?).await?;
-
-        metrics.inbound_req_inc(DaphneRequestType::Upload);
-        Ok(())
-    }
-
-    /// Handle a collect job from the Collector. The response is the URI that the Collector will
-    /// poll later on to get the collection.
-    async fn handle_collect_job_req(&self, req: &DapRequest<S>) -> Result<Url, DapAbort> {
-        let now = self.get_current_time();
-        let metrics = self.metrics().with_host(req.host());
-        let task_id = req.task_id()?;
-        debug!("collect for task {task_id}");
-
-        // Check whether the DAP version indicated by the sender is supported.
-        if req.version == DapVersion::Unknown {
-            return Err(DapAbort::version_unknown());
-        }
-
-        check_request_content_type(req, DapMediaType::CollectReq)?;
-
-        resolve_taskprov(self, task_id, req, None).await?;
-
-        if let Some(reason) = self.unauthorized_reason(req).await? {
-            error!("aborted unauthorized collect request: {reason}");
-            return Err(DapAbort::UnauthorizedRequest {
-                detail: reason,
-                task_id: task_id.clone(),
-            });
-        }
-
-        let mut collect_req =
-            CollectionReq::get_decoded_with_param(&req.version, req.payload.as_ref())
-                .map_err(|e| DapAbort::from_codec_error(e, task_id.clone()))?;
-        let wrapped_task_config = self
-            .get_task_config_for(Cow::Borrowed(req.task_id()?))
-            .await?
-            .ok_or(DapAbort::UnrecognizedTask)?;
-        let task_config = wrapped_task_config.as_ref();
-
-        // Check whether the DAP version in the request matches the task config.
-        if task_config.version != req.version {
-            return Err(DapAbort::version_mismatch(req.version, task_config.version));
-        }
-
-        if collect_req.query == Query::FixedSizeCurrentBatch {
-            // This is where we assign the current batch, and convert the
-            // Query::FixedSizeCurrentBatch into a Query::FixedSizeByBatchId.
-            //
-            // TODO(bhalleycf) Note that currently we are just looking at the
-            // head of the uncollected batch queue, so there is no parallelism
-            // possible for collectors on a given task.  To allow multiple
-            // batches for a task to be collected concurrently for the same task,
-            // we'd need a more complex DO state that allowed us to have batch
-            // state go from unassigned -> in-progress -> complete.
-            let batch_id = self.current_batch(task_id).await?;
-            debug!("FixedSize batch id is {batch_id}");
-            collect_req.query = Query::FixedSizeByBatchId { batch_id };
-        }
-
-        // Ensure the batch boundaries are valid and that the batch doesn't overlap with previosuly
-        // collected batches.
-        let batch_selector = BatchSelector::try_from(collect_req.query.clone())?;
-        check_batch(
-            self,
-            task_config,
-            task_id,
-            &batch_selector,
-            &collect_req.agg_param,
-            now,
-        )
-        .await?;
-
-        // draft02 compatibility: In draft02, the collection job ID is generated as a result of the
-        // initial collection request, whereas in the latest draft, the collection job ID is parsed
-        // from the request path.
-        let collect_job_id = match (req.version, &req.resource) {
-            (DapVersion::Draft02, DapResource::Undefined) => None,
-            (DapVersion::Draft05, DapResource::CollectionJob(ref collect_job_id)) => {
-                Some(collect_job_id.clone())
-            }
-            (DapVersion::Draft05, DapResource::Undefined) => {
-                return Err(DapAbort::BadRequest("undefined resource".into()));
-            }
-            _ => unreachable!("unhandled resource {:?}", req.resource),
-        };
-
-        let collect_job_uri = self
-            .init_collect_job(task_id, &collect_job_id, &collect_req)
-            .await?;
-
-        metrics.inbound_req_inc(DaphneRequestType::Collect);
-        Ok(collect_job_uri)
-    }
-
-    /// Run an aggregation job for a set of reports. Return the number of reports that were
-    /// aggregated successfully.
-    //
-    // TODO Handle non-encodable messages gracefully. The length of `reports` may be too long to
-    // encode in `AggregationJobInitReq`, in which case this method will panic. We should increase
-    // the capacity of this message in the spec. In the meantime, we should at a minimum log this
-    // when it happens.
-    async fn run_agg_job(
-        &self,
-        task_id: &TaskId,
-        task_config: &DapTaskConfig,
-        part_batch_sel: &PartialBatchSelector,
-        reports: Vec<Report>,
-        host: &str,
-    ) -> Result<u64, DapAbort> {
-        let metrics = self.metrics().with_host(host);
-
-        // Prepare AggregationJobInitReq.
-        let agg_job_id = MetaAggregationJobId::gen_for_version(&task_config.version);
-        let transition = task_config
-            .vdaf
-            .produce_agg_job_init_req(
-                self,
-                self,
-                task_id,
-                task_config,
-                &agg_job_id,
-                part_batch_sel,
-                reports,
-                &metrics,
-            )
-            .await?;
-        let (state, agg_job_init_req) = match transition {
-            DapLeaderTransition::Continue(state, agg_job_init_req) => (state, agg_job_init_req),
-            DapLeaderTransition::Skip => return Ok(0),
-            DapLeaderTransition::Uncommitted(..) => {
-                return Err(fatal_error!(err = "unexpected state transition (uncommitted)").into())
-            }
-        };
-        let is_put = task_config.version != DapVersion::Draft02;
-        let url_path = if task_config.version == DapVersion::Draft02 {
-            "aggregate".to_string()
-        } else {
-            format!(
-                "tasks/{}/aggregation_jobs/{}",
-                task_id.to_base64url(),
-                agg_job_id.to_base64url()
-            )
-        };
-
-        // Send AggregationJobInitReq and receive AggregationJobResp.
-        let resp = leader_post!(
-            self,
-            task_id,
-            task_config,
-            &url_path,
-            DapMediaType::AggregationJobInitReq,
-            DapMediaType::AggregationJobResp,
-            agg_job_id.for_request_path(),
-            agg_job_init_req.get_encoded_with_param(&task_config.version),
-            is_put
-        );
-        let agg_job_resp = AggregationJobResp::get_decoded(&resp.payload)
-            .map_err(|e| DapAbort::from_codec_error(e, task_id.clone()))?;
-
-        // Prepare AggreagteContinueReq.
-        let transition = task_config.vdaf.handle_agg_job_resp(
-            task_id,
-            &agg_job_id,
-            state,
-            agg_job_resp,
-            task_config.version,
-            &metrics,
-        )?;
-        let (uncommited, agg_job_cont_req) = match transition {
-            DapLeaderTransition::Uncommitted(uncommited, agg_job_cont_req) => {
-                (uncommited, agg_job_cont_req)
-            }
-            DapLeaderTransition::Skip => return Ok(0),
-            DapLeaderTransition::Continue(..) => {
-                return Err(fatal_error!(err = "unexpected state transition (continue)").into())
-            }
-        };
-
-        // Send AggregationJobContinueReq and receive AggregationJobResp.
-        let resp = leader_post!(
-            self,
-            task_id,
-            task_config,
-            &url_path,
-            DapMediaType::AggregationJobContinueReq,
-            DapMediaType::agg_job_cont_resp_for_version(task_config.version),
-            agg_job_id.for_request_path(),
-            agg_job_cont_req.get_encoded_with_param(&task_config.version),
-            false
-        );
-        let agg_job_resp = AggregationJobResp::get_decoded(&resp.payload)
-            .map_err(|e| DapAbort::from_codec_error(e, task_id.clone()))?;
-
-        // Commit the output shares.
-        let out_shares =
-            task_config
-                .vdaf
-                .handle_final_agg_job_resp(uncommited, agg_job_resp, &metrics)?;
-        let out_shares_count = out_shares.len() as u64;
-        self.put_out_shares(task_id, part_batch_sel, out_shares)
-            .await?;
-
-        metrics.report_inc_by("aggregated", out_shares_count);
-        Ok(out_shares_count)
-    }
-
-    /// Handle a pending collection job. If the results are ready, then compute the aggregate
-    /// results and store them to be retrieved by the Collector later. Returns the number of
-    /// reports in the batch.
-    async fn run_collect_job(
-        &self,
-        task_id: &TaskId,
-        collect_id: &CollectionJobId,
-        task_config: &DapTaskConfig,
-        collect_req: &CollectionReq,
-        host: &str,
-    ) -> Result<u64, DapAbort> {
-        let metrics = self.metrics().with_host(host);
-
-        debug!("collecting id {collect_id}");
-        let batch_selector = BatchSelector::try_from(collect_req.query.clone())?;
-        let leader_agg_share = self.get_agg_share(task_id, &batch_selector).await?;
-
-        // Check the batch size. If not not ready, then return early.
-        //
-        // TODO Consider logging this error, as it should never happen.
-        if !task_config.is_report_count_compatible(task_id, leader_agg_share.report_count)? {
-            return Ok(0);
-        }
-
-        let batch_selector = BatchSelector::try_from(collect_req.query.clone())?;
-
-        // Prepare the Leader's aggregate share.
-        let leader_enc_agg_share = task_config.vdaf.produce_leader_encrypted_agg_share(
-            &task_config.collector_hpke_config,
-            task_id,
-            &batch_selector,
-            &leader_agg_share,
-            task_config.version,
-        )?;
-
-        // Prepare AggregateShareReq.
-        let agg_share_req = AggregateShareReq {
-            draft02_task_id: task_id.for_request_payload(&task_config.version),
-            batch_sel: batch_selector.clone(),
-            agg_param: collect_req.agg_param.clone(),
-            report_count: leader_agg_share.report_count,
-            checksum: leader_agg_share.checksum,
-        };
-
-        let url_path = if task_config.version == DapVersion::Draft02 {
-            "aggregate_share".to_string()
-        } else {
-            format!("tasks/{}/aggregate_shares", task_id.to_base64url())
-        };
-
-        // Send AggregateShareReq and receive AggregateShareResp.
-        let resp = leader_post!(
-            self,
-            task_id,
-            task_config,
-            &url_path,
-            DapMediaType::AggregateShareReq,
-            DapMediaType::AggregateShare,
-            DapResource::Undefined,
-            agg_share_req.get_encoded_with_param(&task_config.version),
-            false
-        );
-        let agg_share_resp = AggregateShare::get_decoded(&resp.payload)
-            .map_err(|e| DapAbort::from_codec_error(e, task_id.clone()))?;
-        // For draft05 and later, the Collection message includes the smallest quantized time
-        // interval containing all reports in the batch.
-        let interval = match task_config.version {
-            DapVersion::Draft02 => None,
-            DapVersion::Draft05 => {
-                let low = task_config.quantized_time_lower_bound(leader_agg_share.min_time);
-                let high = task_config.quantized_time_upper_bound(leader_agg_share.max_time);
-                Some(Interval {
-                    start: low,
-                    duration: if high > low {
-                        high - low
-                    } else {
-                        // This should never happen!
-                        task_config.time_precision
-                    },
-                })
-            }
-            _ => unreachable!("unhandled version {}", task_config.version),
-        };
-
-        // Complete the collect job.
-        let collection = Collection {
-            part_batch_sel: batch_selector.into(),
-            report_count: leader_agg_share.report_count,
-            interval,
-            encrypted_agg_shares: vec![leader_enc_agg_share, agg_share_resp.encrypted_agg_share],
-        };
-        self.finish_collect_job(task_id, collect_id, &collection)
-            .await?;
-
-        // Mark reports as collected.
-        self.mark_collected(task_id, &agg_share_req.batch_sel)
-            .await?;
-
-        metrics.report_inc_by("collected", agg_share_req.report_count);
-        Ok(agg_share_req.report_count)
-    }
-
-    /// Fetch a set of reports grouped by task, then run an aggregation job for each task. Once all
-    /// jobs are completed, process the collect job queue. It is not safe to run multiple instances
-    /// of this function in parallel.
-    ///
-    /// This method is geared primarily towards testing. It also demonstrates how to properly
-    /// synchronize collect and aggregation jobs. If used in a large DAP deployment, it is likely
-    /// create a bottleneck. Such deployments can improve throughput by running many aggregation
-    /// jobs in parallel.
-    async fn process(
-        &self,
-        selector: &Self::ReportSelector,
-        host: &str,
-    ) -> Result<DapLeaderProcessTelemetry, DapAbort> {
-        let mut telem = DapLeaderProcessTelemetry::default();
-
-        // Fetch reports and run an aggregation job for each task.
-        for (task_id, reports) in self.get_reports(selector).await?.into_iter() {
-            let task_config = self
-                .get_task_config_for(Cow::Owned(task_id.clone()))
-                .await?
-                .ok_or(DapAbort::UnrecognizedTask)?;
-
-            for (part_batch_sel, reports) in reports.into_iter() {
-                // TODO Consider splitting reports into smaller chunks.
-                // TODO Consider handling tasks in parallel.
-                telem.reports_processed += reports.len() as u64;
-                debug!(
-                    "process {} reports for task {task_id} with selector {part_batch_sel:?}",
-                    reports.len()
-                );
-                if !reports.is_empty() {
-                    telem.reports_aggregated += self
-                        .run_agg_job(
-                            &task_id,
-                            task_config.as_ref(),
-                            &part_batch_sel,
-                            reports,
-                            host,
-                        )
-                        .await?;
-                }
-            }
-        }
-        // Process pending collect jobs. We wait until all aggregation jobs are finished before
-        // proceeding to this step. This is to prevent a race condition involving an aggregate
-        // share computed during a collect job and any output shares computed during an aggregation
-        // job.
-        for (task_id, collect_id, collect_req) in self.get_pending_collect_jobs().await? {
-            let task_config = self
-                .get_task_config_for(Cow::Owned(task_id.clone()))
-                .await?
-                .ok_or(DapAbort::UnrecognizedTask)?;
-
-            telem.reports_collected += self
-                .run_collect_job(
-                    &task_id,
-                    &collect_id,
-                    task_config.as_ref(),
-                    &collect_req,
-                    host,
-                )
-                .await?;
-        }
-
-        Ok(telem)
-    }
-}
-
-/// DAP Helper functionality.
-#[async_trait(?Send)]
-pub trait DapHelper<S>: DapAggregator<S> {
-    /// Store the Helper's aggregation-flow state unless it already exists. Returns a boolean
-    /// indicating if the operation succeeded.
-    async fn put_helper_state_if_not_exists(
-        &self,
-        task_id: &TaskId,
-        agg_job_id: &MetaAggregationJobId,
-        helper_state: &DapHelperState,
-    ) -> Result<bool, DapError>;
-
-    /// Fetch the Helper's aggregation-flow state. `None` is returned if the Helper has no state
-    /// associated with the given task and aggregation job.
-    async fn get_helper_state(
-        &self,
-        task_id: &TaskId,
-        agg_job_id: &MetaAggregationJobId,
-    ) -> Result<Option<DapHelperState>, DapError>;
-
-    /// Handle a request pertaining to an aggregation job.
-    async fn handle_agg_job_req(&self, req: &DapRequest<S>) -> Result<DapResponse, DapAbort> {
-        let metrics = self.metrics().with_host(req.host());
-        let task_id = req.task_id()?;
-
-        // Check whether the DAP version indicated by the sender is supported.
-        if req.version == DapVersion::Unknown {
-            return Err(DapAbort::version_unknown());
-        }
-
-        if let Some(reason) = self.unauthorized_reason(req).await? {
-            error!("aborted unauthorized collect request: {reason}");
-            return Err(DapAbort::UnauthorizedRequest {
-                detail: reason,
-                task_id: task_id.clone(),
-            });
-        }
-
-        match req.media_type {
-            DapMediaType::AggregationJobInitReq => {
-                let agg_job_init_req =
-                    AggregationJobInitReq::get_decoded_with_param(&req.version, &req.payload)
-                        .map_err(|e| DapAbort::from_codec_error(e, task_id.clone()))?;
-
-                metrics.agg_job_observe_batch_size(agg_job_init_req.report_shares.len());
-
-                // taskprov: Resolve the task config to use for the request. We also need to ensure
-                // that all of the reports include the task config in the report extensions. (See
-                // section 6 of draft-wang-ppm-dap-taskprov-02.)
-                let mut first_metadata: Option<&ReportMetadata> = None;
-                let global_config = self.get_global_config();
-                if global_config.allow_taskprov {
-                    let using_taskprov = agg_job_init_req
-                        .report_shares
-                        .iter()
-                        .filter(|share| {
-                            share
-                                .report_metadata
-                                .is_taskprov(global_config.taskprov_version, task_id)
-                        })
-                        .count();
-
-                    if using_taskprov == agg_job_init_req.report_shares.len() {
-                        // All the extensions use taskprov and look ok, so compute first_metadata.
-                        // Note this will always be Some().
-                        first_metadata = agg_job_init_req
-                            .report_shares
-                            .first()
-                            .map(|report_share| &report_share.report_metadata);
-                    } else if using_taskprov != 0 {
-                        // It's not all taskprov or no taskprov, so it's an error.
-                        return Err(DapAbort::UnrecognizedMessage {
-                            detail: "some reports include the taskprov extensions and some do not"
-                                .to_string(),
-                            task_id: Some(task_id.clone()),
-                        });
-                    }
-                }
-                resolve_taskprov(self, task_id, req, first_metadata).await?;
-
-                let wrapped_task_config = self
-                    .get_task_config_for(Cow::Borrowed(task_id))
-                    .await?
-                    .ok_or(DapAbort::UnrecognizedTask)?;
-                let task_config = wrapped_task_config.as_ref();
-
-                // draft02 compatibility: In draft02, the aggregation job ID is parsed from the
-                // HTTP request payload; in the latest draft, the aggregation job ID is parsed from
-                // the request path.
-                let agg_job_id = match (
-                    req.version,
-                    &req.resource,
-                    &agg_job_init_req.draft02_agg_job_id,
-                ) {
-                    (DapVersion::Draft02, DapResource::Undefined, Some(ref agg_job_id)) => {
-                        MetaAggregationJobId::Draft02(Cow::Borrowed(agg_job_id))
-                    }
-                    (DapVersion::Draft05, DapResource::AggregationJob(ref agg_job_id), None) => {
-                        MetaAggregationJobId::Draft05(Cow::Borrowed(agg_job_id))
-                    }
-                    (DapVersion::Draft05, DapResource::Undefined, None) => {
-                        return Err(DapAbort::BadRequest("undefined resource".into()));
-                    }
-                    _ => unreachable!("unhandled resource {:?}", req.resource),
-                };
-
-                // Check whether the DAP version in the request matches the task config.
-                if task_config.version != req.version {
-                    return Err(DapAbort::version_mismatch(req.version, task_config.version));
-                }
-
-                // Ensure we know which batch the request pertains to.
-                check_part_batch(
-                    task_id,
-                    task_config,
-                    &agg_job_init_req.part_batch_sel,
-                    &agg_job_init_req.agg_param,
-                )?;
-
-                let transition = task_config
-                    .vdaf
-                    .handle_agg_job_init_req(
-                        self,
-                        self,
-                        task_id,
-                        task_config,
-                        &agg_job_init_req,
-                        &metrics,
-                    )
-                    .map_err(DapError::Abort)
-                    .await?;
-
-                let agg_job_resp = match transition {
-                    DapHelperTransition::Continue(state, agg_job_resp) => {
-                        if !self
-                            .put_helper_state_if_not_exists(task_id, &agg_job_id, &state)
-                            .await?
-                        {
-                            // TODO spec: Consider an explicit abort for this case.
-                            return Err(DapAbort::BadRequest(
-                                "unexpected message for aggregation job (already exists)".into(),
-                            ));
-                        }
-                        agg_job_resp
-                    }
-                    DapHelperTransition::Finish(..) => {
-                        return Err(fatal_error!(err = "unexpected transition (finished)").into());
-                    }
-                };
-
-                self.audit_log().on_aggregation_job(
-                    req.host(),
-                    task_id,
-                    task_config,
-                    agg_job_init_req.report_shares.len() as u64,
-                    AggregationJobAuditAction::Init,
-                );
-
-                metrics.agg_job_started_inc();
-                metrics.inbound_req_inc(DaphneRequestType::Aggregate);
-                Ok(DapResponse {
-                    version: req.version,
-                    media_type: DapMediaType::AggregationJobResp,
-                    payload: agg_job_resp.get_encoded(),
-                })
-            }
-            DapMediaType::AggregationJobContinueReq => {
-                resolve_taskprov(self, task_id, req, None).await?;
-                let wrapped_task_config = self
-                    .get_task_config_for(Cow::Borrowed(task_id))
-                    .await?
-                    .ok_or(DapAbort::UnrecognizedTask)?;
-                let task_config = wrapped_task_config.as_ref();
-
-                // Check whether the DAP version in the request matches the task config.
-                if task_config.version != req.version {
-                    return Err(DapAbort::version_mismatch(req.version, task_config.version));
-                }
-
-                let agg_job_cont_req =
-                    AggregationJobContinueReq::get_decoded_with_param(&req.version, &req.payload)
-                        .map_err(|e| DapAbort::from_codec_error(e, task_id.clone()))?;
-
-                // draft02 compatibility: In draft02, the aggregation job ID is parsed from the
-                // HTTP request payload; in the latest, the aggregation job ID is parsed from the
-                // request path.
-                let agg_job_id = match (
-                    req.version,
-                    &req.resource,
-                    &agg_job_cont_req.draft02_agg_job_id,
-                ) {
-                    (DapVersion::Draft02, DapResource::Undefined, Some(ref agg_job_id)) => {
-                        MetaAggregationJobId::Draft02(Cow::Borrowed(agg_job_id))
-                    }
-                    (DapVersion::Draft05, DapResource::AggregationJob(ref agg_job_id), None) => {
-                        MetaAggregationJobId::Draft05(Cow::Borrowed(agg_job_id))
-                    }
-                    (DapVersion::Draft05, DapResource::Undefined, None) => {
-                        return Err(DapAbort::BadRequest("undefined resource".into()));
-                    }
-                    _ => unreachable!("unhandled resource {:?}", req.resource),
-                };
-
-                let state = self.get_helper_state(task_id, &agg_job_id).await?.ok_or(
-                    DapAbort::UnrecognizedAggregationJob {
-                        task_id: task_id.clone(),
-                        agg_job_id_base64url: agg_job_id.to_base64url(),
-                    },
-                )?;
-                let part_batch_sel = state.part_batch_sel.clone();
-                let transition = task_config.vdaf.handle_agg_job_cont_req(
-                    task_id,
-                    &agg_job_id,
-                    state,
-                    &agg_job_cont_req,
-                    &metrics,
-                )?;
-
-                let (agg_job_resp, out_shares_count) = match transition {
-                    DapHelperTransition::Continue(..) => {
-                        return Err(fatal_error!(err = "unexpected transition (continued)").into());
-                    }
-                    DapHelperTransition::Finish(out_shares, agg_job_resp) => {
-                        let out_shares_count = u64::try_from(out_shares.len()).unwrap();
-                        self.put_out_shares(task_id, &part_batch_sel, out_shares)
-                            .await?;
-                        (agg_job_resp, out_shares_count)
-                    }
-                };
-
-                self.audit_log().on_aggregation_job(
-                    req.host(),
-                    task_id,
-                    task_config,
-                    out_shares_count,
-                    AggregationJobAuditAction::Continue,
-                );
-
-                metrics.report_inc_by("aggregated", out_shares_count);
-                metrics.agg_job_completed_inc();
-                metrics.inbound_req_inc(DaphneRequestType::Aggregate);
-                Ok(DapResponse {
-                    version: req.version,
-                    media_type: DapMediaType::agg_job_cont_resp_for_version(task_config.version),
-                    payload: agg_job_resp.get_encoded(),
-                })
-            }
-            //TODO spec: Specify this behavior.
-            _ => Err(DapAbort::BadRequest("unexpected media type".into())),
-        }
-    }
-
-    /// Handle a request for an aggregate share. This is called by the Leader to complete a
-    /// collection job.
-    async fn handle_agg_share_req(&self, req: &DapRequest<S>) -> Result<DapResponse, DapAbort> {
-        let now = self.get_current_time();
-        let metrics = self.metrics().with_host(req.host());
-        let task_id = req.task_id()?;
-
-        // Check whether the DAP version indicated by the sender is supported.
-        if req.version == DapVersion::Unknown {
-            return Err(DapAbort::version_unknown());
-        }
-
-        check_request_content_type(req, DapMediaType::AggregateShareReq)?;
-
-        resolve_taskprov(self, task_id, req, None).await?;
-
-        if let Some(reason) = self.unauthorized_reason(req).await? {
-            error!("aborted unauthorized collect request: {reason}");
-            return Err(DapAbort::UnauthorizedRequest {
-                detail: reason,
-                task_id: task_id.clone(),
-            });
-        }
-
-        let wrapped_task_config = self
-            .get_task_config_for(Cow::Borrowed(req.task_id()?))
-            .await?
-            .ok_or(DapAbort::UnrecognizedTask)?;
-        let task_config = wrapped_task_config.as_ref();
-
-        // Check whether the DAP version in the request matches the task config.
-        if task_config.version != req.version {
-            return Err(DapAbort::version_mismatch(req.version, task_config.version));
-        }
-
-        let agg_share_req = AggregateShareReq::get_decoded_with_param(&req.version, &req.payload)
-            .map_err(|e| DapAbort::from_codec_error(e, task_id.clone()))?;
-
-        // Ensure the batch boundaries are valid and that the batch doesn't overlap with previosuly
-        // collected batches.
-        check_batch(
-            self,
-            task_config,
-            task_id,
-            &agg_share_req.batch_sel,
-            &agg_share_req.agg_param,
-            now,
-        )
-        .await?;
-
-        let agg_share = self
-            .get_agg_share(task_id, &agg_share_req.batch_sel)
-            .await?;
-
-        // Check that we have aggreagted the same set of reports as the Leader.
-        if agg_share_req.report_count != agg_share.report_count
-            || !constant_time_eq(&agg_share_req.checksum, &agg_share.checksum)
-        {
-            return Err(DapAbort::BatchMismatch{
-                detail: format!("Either the report count or checksum does not match: the Leader computed {} and {}; the Helper computed {} and {}.",
-                    agg_share_req.report_count,
-                    hex::encode(agg_share_req.checksum),
-                    agg_share.report_count,
-                    hex::encode(agg_share.checksum)),
-                task_id: task_id.clone(),
-            });
-        }
-
-        // Check the batch size.
-        if !task_config
-            .is_report_count_compatible(task_id, agg_share.report_count)
-            .unwrap_or(false)
-        {
-            return Err(DapAbort::InvalidBatchSize {
-                detail: format!(
-                    "Report count ({}) is less than minimum ({})",
-                    agg_share.report_count, task_config.min_batch_size
-                ),
-                task_id: task_id.clone(),
-            });
-        }
-
-        // Mark each aggregated report as collected.
-        self.mark_collected(task_id, &agg_share_req.batch_sel)
-            .await?;
-
-        let encrypted_agg_share = task_config.vdaf.produce_helper_encrypted_agg_share(
-            &task_config.collector_hpke_config,
-            task_id,
-            &agg_share_req.batch_sel,
-            &agg_share,
-            task_config.version,
-        )?;
-
-        let agg_share_resp = AggregateShare {
-            encrypted_agg_share,
-        };
-
-        metrics.report_inc_by("collected", agg_share_req.report_count);
-        metrics.inbound_req_inc(DaphneRequestType::Collect);
-        Ok(DapResponse {
-            version: req.version,
-            media_type: DapMediaType::AggregateShare,
-            payload: agg_share_resp.get_encoded(),
-        })
-    }
-}
-
-fn check_part_batch(
-    task_id: &TaskId,
-    task_config: &DapTaskConfig,
-    part_batch_sel: &PartialBatchSelector,
-    agg_param: &[u8],
-) -> Result<(), DapAbort> {
-    if !task_config.query.is_valid_part_batch_sel(part_batch_sel) {
-        return Err(DapAbort::query_mismatch(
-            task_id,
-            &task_config.query,
-            part_batch_sel,
-        ));
-    }
-
-    // Check that the aggregation parameter is suitable for the given VDAF.
-    if !task_config.vdaf.is_valid_agg_param(agg_param) {
-        // TODO spec: Define this behavior.
-        return Err(DapAbort::UnrecognizedMessage {
-            detail: "invalid aggregation parameter".into(),
-            task_id: Some(task_id.clone()),
-        });
-    }
-
-    Ok(())
-}
-
-async fn check_batch<S>(
-    agg: &impl DapAggregator<S>,
-    task_config: &DapTaskConfig,
-    task_id: &TaskId,
-    batch_sel: &BatchSelector,
-    agg_param: &[u8],
-    now: Time,
-) -> Result<(), DapAbort> {
-    let global_config = agg.get_global_config();
-    let batch_overlapping = agg.is_batch_overlapping(task_id, batch_sel);
-
-    // Check that the aggregation parameter is suitable for the given VDAF.
-    if !task_config.vdaf.is_valid_agg_param(agg_param) {
-        // TODO spec: Define this behavior.
-        return Err(DapAbort::UnrecognizedMessage {
-            detail: "invalid aggregation parameter".into(),
-            task_id: Some(task_id.clone()),
-        });
-    }
-
-    // Check that the batch boundaries are valid.
-    match (&task_config.query, batch_sel) {
-        (DapQueryConfig::TimeInterval { .. }, BatchSelector::TimeInterval { batch_interval }) => {
-            if batch_interval.start % task_config.time_precision != 0
-                || batch_interval.duration % task_config.time_precision != 0
-                || batch_interval.duration < task_config.time_precision
-            {
-                return Err(DapAbort::BatchInvalid {
-                    detail: format!("The queried batch interval ({batch_interval:?}) is too small or its boundaries are misaligned. The time precision for this task is {}s.", task_config.time_precision),
-                    task_id: task_id.clone(),
-                });
-            }
-
-            if batch_interval.duration > global_config.max_batch_duration {
-                return Err(DapAbort::BadRequest("batch interval too large".to_string()));
-            }
-
-            if now.abs_diff(batch_interval.start) > global_config.min_batch_interval_start {
-                return Err(DapAbort::BadRequest(
-                    "batch interval too far into past".to_string(),
-                ));
-            }
-
-            if now.abs_diff(batch_interval.end()) > global_config.max_batch_interval_end {
-                return Err(DapAbort::BadRequest(
-                    "batch interval too far into future".to_string(),
-                ));
-            }
-        }
-        (DapQueryConfig::FixedSize { .. }, BatchSelector::FixedSizeByBatchId { batch_id }) => {
-            // TODO(cjpatton) The Helper can avoid this callback by first fetching the aggregate
-            // share and aborting with "batchInvalid" if the report count is 0. Depending on how we
-            // resolve https://github.com/ietf-wg-ppm/draft-ietf-ppm-dap/issues/342, this check may
-            // become unnecessary for the Leader.
-            //
-            // Consider removing this callback once we resolve DAP issue #342.
-            if !agg.batch_exists(task_id, batch_id).await? {
-                return Err(DapAbort::BatchInvalid {
-                    detail: format!(
-                        "The queried batch ({}) does not exist.",
-                        batch_id.to_base64url()
-                    ),
-                    task_id: task_id.clone(),
-                });
-            }
-        }
-        _ => {
-            return Err(DapAbort::query_mismatch(
-                task_id,
-                &task_config.query,
-                batch_sel,
-            ))
-        }
-    };
-
-    // Check that the batch does not overlap with any previously collected batch.
-    if batch_overlapping.await? {
-        return Err(DapAbort::batch_overlap(task_id, batch_sel));
-    }
-
-    Ok(())
-}
-
-/// Check for transition failures due to:
-///
-/// * the report having already been processed
-/// * the report having already been collected
-/// * the report not being within time bounds
-///
-/// Returns `Some(TransitionFailure)` if there is a problem,
-/// or `None` if no transition failure occurred.
-pub fn early_metadata_check(
-    metadata: &ReportMetadata,
-    processed: bool,
-    collected: bool,
-    min_time: u64,
-    max_time: u64,
-) -> Option<TransitionFailure> {
-    if processed {
-        Some(TransitionFailure::ReportReplayed)
-    } else if collected {
-        Some(TransitionFailure::BatchCollected)
-    } else if metadata.time < min_time {
-        Some(TransitionFailure::ReportDropped)
-    } else if metadata.time > max_time {
-        Some(TransitionFailure::ReportTooEarly)
-    } else {
-        None
-    }
-}
-
-fn check_response_content_type(resp: &DapResponse, expected: DapMediaType) -> Result<(), DapError> {
-    let want_str = expected
-        .as_str_for_version(resp.version)
-        .expect("could not determine string representation for expected content-type");
-
-    if resp.media_type != expected {
-        if let Some(got_str) = resp.media_type.as_str_for_version(resp.version) {
-            Err(fatal_error!(
-                err = "response from peer has unexpected content-type",
-                got = got_str,
-                want = want_str,
-            ))
-        } else {
-            Err(fatal_error!(
-                err = "response from peer has no content-type",
-                expected = want_str,
-            ))
-        }
-    } else {
-        Ok(())
-    }
-}
-
-fn check_request_content_type<S>(
-    req: &DapRequest<S>,
-    expected: DapMediaType,
-) -> Result<(), DapAbort> {
-    if req.media_type != expected {
-        Err(DapAbort::content_type(req, expected))
-    } else {
-        Ok(())
-    }
-}
-
-async fn resolve_taskprov<S>(
-    agg: &impl DapAggregator<S>,
-    task_id: &TaskId,
-    req: &DapRequest<S>,
-    report_metadata_advertisement: Option<&ReportMetadata>,
-) -> Result<(), DapError> {
-    if agg
-        .get_task_config_for(Cow::Borrowed(task_id))
-        .await?
-        .is_some()
-    {
-        // Task already configured, so nothing to do.
-        return Ok(());
-    }
-
-    let global_config = agg.get_global_config();
-    if !global_config.allow_taskprov {
-        // Taskprov is disabled, so nothing to do.
-        return Ok(());
-    }
-
-    let Some(vdaf_verify_key_init) = agg.taskprov_vdaf_verify_key_init() else {
-        warn!("Taskprov disabled due to missing VDAF verification key initializer.");
-        return Ok(());
-    };
-
-    let Some(collector_hpke_config) = agg.taskprov_collector_hpke_config() else {
-        warn!("Taskprov disabled due to missing Collector HPKE configuration.");
-        return Ok(());
-    };
-
-    let Some(task_config) = taskprov::resolve_advertised_task_config(
-            req,
-            global_config.taskprov_version,
-            vdaf_verify_key_init,
-            collector_hpke_config,
-            task_id,
-            report_metadata_advertisement,
-        )? else {
-            // No task configuration advertised, so nothing to do.
-            return Ok(());
-        };
-
-    // This is the opt-in / opt-out decision point.
-    if let Some(reason) = agg.taskprov_opt_out_reason(&task_config)? {
-        return Err(DapError::Abort(DapAbort::InvalidTask {
-            detail: reason,
-            task_id: task_id.clone(),
-        }));
-    }
-
-    agg.taskprov_put(req, task_config).await?;
-    Ok(())
-}
